# internal/controller/handler/v1/campaigns/claims/gen/server.gen.go
// Package gen provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.8.0 DO NOT EDIT.
package gen

import (
	"bytes"
	"context"
	"encoding/json"
	"fmt"
	"net/http"

	"github.com/labstack/echo/v5"
	"github.com/oapi-codegen/runtime"
)

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// PostCampaignsClaims コードによるクーポンの受け取り
	// (POST /v1/campaigns/claims)
	PostCampaignsClaims(ctx *echo.Context, params PostCampaignsClaimsParams) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// PostCampaignsClaims converts echo context to params.
func (w *ServerInterfaceWrapper) PostCampaignsClaims(ctx *echo.Context) error {
	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params PostCampaignsClaimsParams

	headers := ctx.Request().Header
	// ------------- Required header parameter "Idempotency-Key" -------------
	if valueList, found := headers[http.CanonicalHeaderKey("Idempotency-Key")]; found {
		var IdempotencyKey IdempotencyKeyRequiredParam
		n := len(valueList)
		if n != 1 {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Expected one value for Idempotency-Key, got %d", n))
		}

		err = runtime.BindStyledParameterWithOptions("simple", "Idempotency-Key", valueList[0], &IdempotencyKey, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationHeader, Explode: false, Required: true, Type: "string", Format: ""})
		if err != nil {
			return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter Idempotency-Key: %s", err))
		}

		params.IdempotencyKey = IdempotencyKey
	} else {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Header parameter Idempotency-Key is required, but not found"))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.PostCampaignsClaims(ctx, params)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) echo.RouteInfo
}

// RegisterHandlersOptions configures RegisterHandlersWithOptions.
type RegisterHandlersOptions struct {
	// BaseURL is prepended to every registered path so the API can be served
	// under a prefix.
	BaseURL string
	// OperationMiddlewares lets the caller attach per-operation middleware at
	// registration time. The map key is the OpenAPI `operationId` value as it
	// appears in the spec (the raw, un-normalized form). Operations that have
	// no entry are registered with no extra middleware. A nil map disables
	// per-operation middleware entirely.
	OperationMiddlewares map[string][]echo.MiddlewareFunc
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{})
}

// RegisterHandlersWithBaseURL registers handlers and prepends BaseURL to the
// paths so the API can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {
	RegisterHandlersWithOptions(router, si, RegisterHandlersOptions{BaseURL: baseURL})
}

// RegisterHandlersWithOptions registers handlers using the supplied options,
// including any per-operation middleware.
func RegisterHandlersWithOptions(router EchoRouter, si ServerInterface, options RegisterHandlersOptions) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(options.BaseURL+"/v1/campaigns/claims", wrapper.PostCampaignsClaims, options.OperationMiddlewares["PostCampaignsClaims"]...)

}

type BadRequest400JSONResponse ErrorResponse

type Conflict409JSONResponse ErrorResponse

type InternalServerError500JSONResponse ErrorResponse

type MethodNotAllowed405JSONResponse ErrorResponse

type ServiceUnavailable503JSONResponse ErrorResponse

type Unauthorized401JSONResponse ErrorResponse

type UnprocessableEntity422JSONResponse ErrorResponseWithDetails

type PostCampaignsClaimsRequestObject struct {
	Params PostCampaignsClaimsParams
	Body   *PostCampaignsClaimsJSONRequestBody
}

type PostCampaignsClaimsResponseObject interface {
	VisitPostCampaignsClaimsResponse(w http.ResponseWriter) error
}

type PostCampaignsClaims201JSONResponse CouponResponse

func (response PostCampaignsClaims201JSONResponse) VisitPostCampaignsClaimsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)
	_, err := buf.WriteTo(w)
	return err
}

type PostCampaignsClaims400JSONResponse struct{ BadRequest400JSONResponse }

func (response PostCampaignsClaims400JSONResponse) VisitPostCampaignsClaimsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)
	_, err := buf.WriteTo(w)
	return err
}

type PostCampaignsClaims401JSONResponse struct{ Unauthorized401JSONResponse }

func (response PostCampaignsClaims401JSONResponse) VisitPostCampaignsClaimsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(401)
	_, err := buf.WriteTo(w)
	return err
}

type PostCampaignsClaims405JSONResponse struct {
	MethodNotAllowed405JSONResponse
}

func (response PostCampaignsClaims405JSONResponse) VisitPostCampaignsClaimsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(405)
	_, err := buf.WriteTo(w)
	return err
}

type PostCampaignsClaims409JSONResponse struct{ Conflict409JSONResponse }

func (response PostCampaignsClaims409JSONResponse) VisitPostCampaignsClaimsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)
	_, err := buf.WriteTo(w)
	return err
}

type PostCampaignsClaims422JSONResponse struct {
	UnprocessableEntity422JSONResponse
}

func (response PostCampaignsClaims422JSONResponse) VisitPostCampaignsClaimsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(422)
	_, err := buf.WriteTo(w)
	return err
}

type PostCampaignsClaims500JSONResponse struct {
	InternalServerError500JSONResponse
}

func (response PostCampaignsClaims500JSONResponse) VisitPostCampaignsClaimsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)
	_, err := buf.WriteTo(w)
	return err
}

type PostCampaignsClaims503JSONResponse struct {
	ServiceUnavailable503JSONResponse
}

func (response PostCampaignsClaims503JSONResponse) VisitPostCampaignsClaimsResponse(w http.ResponseWriter) error {

	var buf bytes.Buffer
	if err := json.NewEncoder(&buf).Encode(response); err != nil {
		return err
	}
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(503)
	_, err := buf.WriteTo(w)
	return err
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// PostCampaignsClaims コードによるクーポンの受け取り
	// (POST /v1/campaigns/claims)
	PostCampaignsClaims(ctx context.Context, request PostCampaignsClaimsRequestObject) (PostCampaignsClaimsResponseObject, error)
}

type StrictHandlerFunc func(ctx *echo.Context, request any) (any, error)
type StrictMiddlewareFunc func(f StrictHandlerFunc, operationID string) StrictHandlerFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// PostCampaignsClaims operation middleware
func (sh *strictHandler) PostCampaignsClaims(ctx *echo.Context, params PostCampaignsClaimsParams) error {
	var request PostCampaignsClaimsRequestObject

	request.Params = params

	var body PostCampaignsClaimsJSONRequestBody
	var err error
	if _, ok := ctx.Echo().Binder.(*echo.DefaultBinder); ok {
		// Bind only the request body, so that path and query parameters
		// are not also bound into the body struct.
		err = echo.BindBody(ctx, &body)
	} else {
		// A custom binder is installed on the Echo instance; defer to it
		// entirely, since echo.Binder does not expose body-only binding.
		err = ctx.Bind(&body)
	}
	if err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx *echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.PostCampaignsClaims(ctx.Request().Context(), request.(PostCampaignsClaimsRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "PostCampaignsClaims")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(PostCampaignsClaimsResponseObject); ok {
		return validResponse.VisitPostCampaignsClaimsResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}