# openapi/openapi.gen.yaml
openapi: 3.0.3
info:
  title: Go Boilerplate API
  termsOfService: https://github.com/Tomy-ch/go-boilerplate-echo-oapi-sqlc
  contact:
    name: TomyC258460
    url: https://x.com/TomyC258460
  version: 2.3.0
  description: |
    Reference specification for the sample API (users, products, purchases). It exercises every layer
    of the Onion Architecture end to end: cursor and offset pagination, creation and partial update,
    soft delete, idempotency keys, and role-based authorization.

    Handlers (oapi-codegen) and the published reference documentation are both generated from this
    file, so every endpoint change starts here.
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
    x-internal: true
  - url: https://api.example.com
    description: Production server (replace with actual URL)
    x-internal: false
  - url: https://staging.api.example.com
    description: Staging server (replace with actual URL)
    x-internal: false
  - url: https://development.api.example.com
    description: Development server (replace with actual URL)
    x-internal: true
security: []
...
paths:
...
  /v1/campaigns/claims:
    post:
      operationId: PostCampaignsClaims
      tags:
        - v1/campaigns/claims
      summary: コードによるクーポンの受け取り
      description: |
        キャンペーンコードを示して、クーポンを 1 枚受け取ります。認証必須で、受給者は認証主体自身です。
        受け取ったクーポンは未使用の状態で作られ、GET /v1/users/me/coupons に並びます。
        他の発行経路のクーポンと区別されません。
        **受け取れない理由は区別しません。** 存在しないコード・配布期間の外・停止済み・総枚数上限に到達・
        その利用者が 1 人あたり上限に到達、のいずれも `422` + `details: ["code"]` を返します。
        区別できると、有効なコードを言い当てる手がかりになるためです。形式が不正なコードも同じ応答です。
        次にすべきことはどの場合も同じ（別のコードを使うか諦める）なので、区別する必要もありません。
        **総枚数上限と 1 人あたり上限は、同時に受け取りが殺到しても超えません。** キャンペーンの行を
        先に押さえてから枚数を数えるため、判定と発行の間に他の受け取りが割り込みません。
        その代わり、同じコードへの同時要求は順番に処理されます。
        **Idempotency-Key は必須です**（欠くと 400）。受け取りは取り消せず、二度目の実行は
        もう 1 枚配るためです。同一キーの再送は初回の結果をそのまま返します。
        **別のキーで押し直した場合は、上限に達していなければ改めて 1 枚配られます。**
        本 op は DB の SELECT / INSERT / UPDATE のみで外部依存を持ちませんが、認証段（外部 IdP の
        JWKS 参照）の一時障害で応答不能となり得るため、認証を伴う op の先例に倣い 503 を宣言します。
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKeyRequiredParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CampaignClaimsPostRequest'
      responses:
        '201':
          description: クーポンの受け取りに成功しました
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CouponResponse'
        '400':
          $ref: '#/components/responses/BadRequest400'
        '401':
          $ref: '#/components/responses/Unauthorized401'
        '405':
          $ref: '#/components/responses/MethodNotAllowed405'
        '409':
          $ref: '#/components/responses/Conflict409'
        '422':
          $ref: '#/components/responses/UnprocessableEntity422'
        '500':
          $ref: '#/components/responses/InternalServerError500'
        '503':
          $ref: '#/components/responses/ServiceUnavailable503'
...
components:
  parameters:
...
    IdempotencyKeyRequiredParam:
      name: Idempotency-Key
      in: header
      description: |
        冪等キー。同一 (認証主体, キー) の再送を初回結果のリプレイとして扱う。
        このヘッダを宣言する操作は、ハンドラが必ず idempotency.Run 経由で処理する契約（完全性テストで機械検証）。

        **この操作では必須です。** 他の非冪等操作は再実行しても状態が収束するか（自然キーの upsert）、
        対象の状態が二度目を弾く（廃番済みの商品は再度廃番にならない）ため任意で足りますが、この操作は
        配布の記録を持たず、二度目の実行が一度目と同じだけの新しいクーポンを配ります。取り消す経路も無いため、
        再送を安全にする手段がキーしかありません。
      required: true
      schema:
        type: string
        pattern: ^[\x21-\x7E]+$
        minLength: 1
        maxLength: 255
      example: 3f0c2d21-6c1e-4a2f-9f8b-4b1e7f0a2c11
...
  securitySchemes:
    BearerAuth:
      type: http
      description: |
        認証にBearerトークンを使用することを示します。
        JWT の検証は RFC8725（JSON Web Token Best Current Practices）に従います。
        署名アルゴリズムは許可リストで固定し、`alg: none` と鍵の混同を拒否します。
      scheme: bearer
      bearerFormat: JWT
    BearerAuthRegistration:
      type: http
      description: |
        内部ユーザーがまだ存在しない主体を受け付ける Bearer トークン認証です。
        トークンの検証は BearerAuth と同一で、JWT の検証は RFC8725（JSON Web Token Best Current
        Practices）に従います。署名アルゴリズムは許可リストで固定し、`alg: none` と鍵の混同を拒否します。
        BearerAuth との違いは、外部アイデンティティ（issuer + subject）を内部ユーザーへ解決しない点だけです。

        新規登録は「認証は済んでいるが内部ユーザーはまだ無い」状態から始まります。BearerAuth は
        解決できない主体を 401 で拒むため、登録の入口をそれで守ると登録が永久に成立しません。
        このスキームを宣言してよいのは、内部ユーザーを自ら作る操作だけです。
      scheme: bearer
      bearerFormat: JWT
    BasicAuth:
      type: http
      description: 認証にBasic認証を使用することを示します。
      scheme: basic
    StreamTicket:
      type: apiKey
      in: query
      name: ticket
      description: |
        SSE 接続の認証に用いる不透明な stream ticket（ADR-0074 (query-ticket-stream-authentication)）。
        EventSource はリクエストヘッダを設定できないため query に載せます。
        feature の認可を通った後に発行され、subject / destination / scope / expiry に束縛されます。
        発行から 5 分間は新しい接続に再利用でき、期限後・失効後の新規接続は 401 で拒否されます。
        サーバは hash だけを保存し、生値を log / trace / error に出しません。
  schemas:
    ErrorResponse:
      type: object
      description: |
        エラーレスポンスの共通スキーマ（base）。 details は返さない。details を返すエンドポイントは ErrorResponseWithDetails を参照する。
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: 機械的に処理可能なエラーコード
          example: EXAMPLE_ERROR
        message:
          type: string
          description: 人間が読めるエラーメッセージ
          example: 定義されたエラーメッセージ
        requestId:
          type: string
          description: リクエストID
          example: abcdefghijklmnopqrstuvwxyz
...
    ErrorResponseWithDetails:
      type: object
      description: |
        詳細識別子付きエラーレスポンススキーマ。 この schema を error レスポンスに宣言したエンドポイントだけが details を返せる（opt-in）。 未宣言のエンドポイントでは errorhandler が details を fail-closed で落とす。
      required:
        - code
        - message
        - requestId
      properties:
        code:
          type: string
          description: 機械的に処理可能なエラーコード
          example: EXAMPLE_ERROR
        message:
          type: string
          description: 人間が読めるエラーメッセージ
          example: 定義されたエラーメッセージ
        requestId:
          type: string
          description: リクエストID
          example: abcdefghijklmnopqrstuvwxyz
        details:
          type: array
          description: エラー詳細を示す公開可能な識別子(任意)。検証エラーでは不正なフィールド名が入る
          items:
            type: string
          example:
            - firstName
            - email
...
    CouponDiscount:
      type: object
      description: |
        クーポンの値引き。いくら引くかを表します。適用範囲（どの明細が対象か）とは独立した軸です。
      required:
        - kind
        - value
        - maxAmount
      properties:
        kind:
          type: string
          enum:
            - flat
            - rate
          description: |
            値引きの決まり方。`flat` は金額を直接差し引く定額、`rate` は対象額に率を掛ける定率です。
          example: rate
        value:
          type: string
          pattern: ^\d+(\.\d+)?$
          description: |
            種別における値。定額なら差し引く金額、定率なら対象額に掛ける率を、
            正確な十進量を保つ decimal 文字列で表します（例 `"0.10"` は 10% 引き）。
            JSON number は IEEE754 double として復元され精度を失うため、文字列で表現します。
          example: '0.10'
        maxAmount:
          type: integer
          format: int64
          nullable: true
          description: |
            定率の値引きが 1 回に引ける額の上限。USD セント単位の整数です。
            定率（`rate`）にのみ意味を持ち、定額（`flat`）では常に null です。
            上限が無い場合も null で、そのときは対象額に率を掛けた額がそのまま引かれます。
          example: 2000
    CouponScope:
      type: object
      description: |
        クーポンの適用範囲。どの明細が対象かを表します。値引き（いくら引くか）とは独立した軸です。
      required:
        - kind
        - targetId
      properties:
        kind:
          type: string
          enum:
            - all
            - category
            - product
          description: |
            適用範囲の決まり方。`all` は対象を絞らない全体、`category` は特定の商品カテゴリに属する明細だけ、
            `product` は特定の商品の明細だけを対象とします。
          example: category
        targetId:
          type: string
          format: uuid
          nullable: true
          description: |
            範囲を絞る対象の識別子。`category` なら商品カテゴリID、`product` なら商品IDです。
            `all` では対象を持たないため null になります。
          example: 5dd52d84-78eb-4a52-ba0b-2e11c95c2af2
    CouponResponse:
      type: object
      description: |
        クーポン。値引きと適用範囲という直交する 2 つの軸を持ちます。
        1 枚のクーポンは 1 つの値引きと 1 つの適用範囲を持ち、複数枚の併用も複数範囲の合成も表しません。
      required:
        - id
        - discount
        - scope
        - expiresAt
        - usedAt
        - issuedAt
        - minPurchaseAmount
        - usableFrom
      properties:
        id:
          type: string
          format: uuid
          description: クーポンID
          example: 0193a1c0-0001-7000-8000-000000000002
        discount:
          $ref: '#/components/schemas/CouponDiscount'
        scope:
          $ref: '#/components/schemas/CouponScope'
        expiresAt:
          type: string
          format: date-time
          description: |
            有効期限。この時刻ちょうどを含めて以降は失効として扱います。
            失効を一括更新する機構は持たず、判定のたびに現在時刻と突き合わせます。
          example: '2026-10-01T00:00:00Z'
        usedAt:
          type: string
          format: date-time
          nullable: true
          description: 使用日時。未使用の場合は null です。使用済みへの遷移は一度きりで取り消せません。
          example: null
        issuedAt:
          type: string
          format: date-time
          description: 発行日時
          example: '2026-09-01T00:00:00Z'
        minPurchaseAmount:
          type: integer
          format: int64
          nullable: true
          description: |
            このクーポンを使うために必要な購入額の下限。USD セント単位の整数です。
            条件が無い場合は null で、そのときは購入額を問わず使えます。
          example: 5000
        usableFrom:
          type: string
          format: date-time
          nullable: true
          description: |
            使えるようになる日時。この時刻ちょうどを含めて以降に使えます。
            条件が無い場合は null で、そのときは発行した時点から使えます。
          example: null
...
    CampaignClaimsPostRequest:
      type: object
      description: |
        コードの提示による受け取りのリクエスト（application/json）。認証必須で、
        認証主体自身がクーポンを 1 枚受け取ります。
        コードはボディで受け取ります。URL に置くとアクセスログや Referer に残り、
        コードそのものが受け取る資格である以上、それは資格情報を記録することになるためです。
      additionalProperties: false
      required:
        - code
      properties:
        code:
          type: string
          description: キャンペーンコード。前後の空白と大文字・小文字の違いは無視されます。 受け取れない場合の理由（存在しない・配布期間外・停止済み・上限到達）は区別されません。
          example: WELCOME-2026
...
  responses:
    MethodNotAllowed405:
      description: 許可されていないリクエスト方法です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadRequest400:
      description: リクエストが不正です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized401:
      description: 認証が必要です。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError500:
      description: サーバーエラーが発生しました。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable503:
      description: サービスが一時的に利用できません。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    Conflict409:
      description: リソースが競合しています。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnprocessableEntity422:
      description: 入力値が要件を満たしていません。details に不正なフィールド名が入ります。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponseWithDetails'
...